+++ esp-hal/src/soc/esp32h2/efuse/fields.rs
@@ -1,412 +1,249 @@
-//! eFuse fields for the ESP32-H2.
-//!
 //! This file was automatically generated, please do not edit it manually!
 //!
-//! For information on how to regenerate these files, please refer to the
-//! `xtask` package's `README.md` file.
-//!
-//! Generated on:   2024-03-11
-//! ESP-IDF Commit: 0de2912f
+//! Generated: 2025-04-22 11:33
+//! Version:   44563d2af4ebdba4db6c0a34a50c94f9
 
-use super::EfuseBlock;
-use crate::soc::efuse_field::EfuseField;
+#![allow(clippy::empty_docs)]
 
-/// `[]` Disable programming of individual eFuses
-pub const WR_DIS: EfuseField = EfuseField::new(EfuseBlock::Block0, 0, 32);
-/// `[]` wr_dis of RD_DIS
-pub const WR_DIS_RD_DIS: EfuseField = EfuseField::new(EfuseBlock::Block0, 0, 1);
-/// `[]` wr_dis of DIS_ICACHE
-pub const WR_DIS_DIS_ICACHE: EfuseField = EfuseField::new(EfuseBlock::Block0, 2, 1);
-/// `[]` wr_dis of DIS_USB_JTAG
-pub const WR_DIS_DIS_USB_JTAG: EfuseField = EfuseField::new(EfuseBlock::Block0, 2, 1);
-/// `[]` wr_dis of POWERGLITCH_EN
-pub const WR_DIS_POWERGLITCH_EN: EfuseField = EfuseField::new(EfuseBlock::Block0, 2, 1);
-/// `[]` wr_dis of DIS_FORCE_DOWNLOAD
-pub const WR_DIS_DIS_FORCE_DOWNLOAD: EfuseField = EfuseField::new(EfuseBlock::Block0, 2, 1);
-/// `[]` wr_dis of SPI_DOWNLOAD_MSPI_DIS
-pub const WR_DIS_SPI_DOWNLOAD_MSPI_DIS: EfuseField = EfuseField::new(EfuseBlock::Block0, 2, 1);
-/// `[WR_DIS.DIS_CAN]` wr_dis of DIS_TWAI
-pub const WR_DIS_DIS_TWAI: EfuseField = EfuseField::new(EfuseBlock::Block0, 2, 1);
-/// `[]` wr_dis of JTAG_SEL_ENABLE
-pub const WR_DIS_JTAG_SEL_ENABLE: EfuseField = EfuseField::new(EfuseBlock::Block0, 2, 1);
-/// `[]` wr_dis of DIS_PAD_JTAG
-pub const WR_DIS_DIS_PAD_JTAG: EfuseField = EfuseField::new(EfuseBlock::Block0, 2, 1);
-/// `[]` wr_dis of DIS_DOWNLOAD_MANUAL_ENCRYPT
-pub const WR_DIS_DIS_DOWNLOAD_MANUAL_ENCRYPT: EfuseField =
-    EfuseField::new(EfuseBlock::Block0, 2, 1);
-/// `[]` wr_dis of WDT_DELAY_SEL
-pub const WR_DIS_WDT_DELAY_SEL: EfuseField = EfuseField::new(EfuseBlock::Block0, 3, 1);
-/// `[]` wr_dis of SPI_BOOT_CRYPT_CNT
-pub const WR_DIS_SPI_BOOT_CRYPT_CNT: EfuseField = EfuseField::new(EfuseBlock::Block0, 4, 1);
-/// `[]` wr_dis of SECURE_BOOT_KEY_REVOKE0
-pub const WR_DIS_SECURE_BOOT_KEY_REVOKE0: EfuseField = EfuseField::new(EfuseBlock::Block0, 5, 1);
-/// `[]` wr_dis of SECURE_BOOT_KEY_REVOKE1
-pub const WR_DIS_SECURE_BOOT_KEY_REVOKE1: EfuseField = EfuseField::new(EfuseBlock::Block0, 6, 1);
-/// `[]` wr_dis of SECURE_BOOT_KEY_REVOKE2
-pub const WR_DIS_SECURE_BOOT_KEY_REVOKE2: EfuseField = EfuseField::new(EfuseBlock::Block0, 7, 1);
-/// `[WR_DIS.KEY0_PURPOSE]` wr_dis of KEY_PURPOSE_0
-pub const WR_DIS_KEY_PURPOSE_0: EfuseField = EfuseField::new(EfuseBlock::Block0, 8, 1);
-/// `[WR_DIS.KEY1_PURPOSE]` wr_dis of KEY_PURPOSE_1
-pub const WR_DIS_KEY_PURPOSE_1: EfuseField = EfuseField::new(EfuseBlock::Block0, 9, 1);
-/// `[WR_DIS.KEY2_PURPOSE]` wr_dis of KEY_PURPOSE_2
-pub const WR_DIS_KEY_PURPOSE_2: EfuseField = EfuseField::new(EfuseBlock::Block0, 10, 1);
-/// `[WR_DIS.KEY3_PURPOSE]` wr_dis of KEY_PURPOSE_3
-pub const WR_DIS_KEY_PURPOSE_3: EfuseField = EfuseField::new(EfuseBlock::Block0, 11, 1);
-/// `[WR_DIS.KEY4_PURPOSE]` wr_dis of KEY_PURPOSE_4
-pub const WR_DIS_KEY_PURPOSE_4: EfuseField = EfuseField::new(EfuseBlock::Block0, 12, 1);
-/// `[WR_DIS.KEY5_PURPOSE]` wr_dis of KEY_PURPOSE_5
-pub const WR_DIS_KEY_PURPOSE_5: EfuseField = EfuseField::new(EfuseBlock::Block0, 13, 1);
-/// `[]` wr_dis of SEC_DPA_LEVEL
-pub const WR_DIS_SEC_DPA_LEVEL: EfuseField = EfuseField::new(EfuseBlock::Block0, 14, 1);
-/// `[]` wr_dis of CRYPT_DPA_ENABLE
-pub const WR_DIS_CRYPT_DPA_ENABLE: EfuseField = EfuseField::new(EfuseBlock::Block0, 14, 1);
-/// `[]` wr_dis of SECURE_BOOT_EN
-pub const WR_DIS_SECURE_BOOT_EN: EfuseField = EfuseField::new(EfuseBlock::Block0, 15, 1);
-/// `[]` wr_dis of SECURE_BOOT_AGGRESSIVE_REVOKE
-pub const WR_DIS_SECURE_BOOT_AGGRESSIVE_REVOKE: EfuseField =
-    EfuseField::new(EfuseBlock::Block0, 16, 1);
-/// `[]` wr_dis of ECDSA_FORCE_USE_HARDWARE_K
-pub const WR_DIS_ECDSA_FORCE_USE_HARDWARE_K: EfuseField =
-    EfuseField::new(EfuseBlock::Block0, 17, 1);
-/// `[]` wr_dis of FLASH_TPUW
-pub const WR_DIS_FLASH_TPUW: EfuseField = EfuseField::new(EfuseBlock::Block0, 18, 1);
-/// `[]` wr_dis of DIS_DOWNLOAD_MODE
-pub const WR_DIS_DIS_DOWNLOAD_MODE: EfuseField = EfuseField::new(EfuseBlock::Block0, 18, 1);
-/// `[]` wr_dis of DIS_DIRECT_BOOT
-pub const WR_DIS_DIS_DIRECT_BOOT: EfuseField = EfuseField::new(EfuseBlock::Block0, 18, 1);
-/// `[WR_DIS.DIS_USB_PRINT]` wr_dis of DIS_USB_SERIAL_JTAG_ROM_PRINT
-pub const WR_DIS_DIS_USB_SERIAL_JTAG_ROM_PRINT: EfuseField =
-    EfuseField::new(EfuseBlock::Block0, 18, 1);
-/// `[]` wr_dis of DIS_USB_SERIAL_JTAG_DOWNLOAD_MODE
-pub const WR_DIS_DIS_USB_SERIAL_JTAG_DOWNLOAD_MODE: EfuseField =
-    EfuseField::new(EfuseBlock::Block0, 18, 1);
-/// `[]` wr_dis of ENABLE_SECURITY_DOWNLOAD
-pub const WR_DIS_ENABLE_SECURITY_DOWNLOAD: EfuseField = EfuseField::new(EfuseBlock::Block0, 18, 1);
-/// `[]` wr_dis of UART_PRINT_CONTROL
-pub const WR_DIS_UART_PRINT_CONTROL: EfuseField = EfuseField::new(EfuseBlock::Block0, 18, 1);
-/// `[]` wr_dis of FORCE_SEND_RESUME
-pub const WR_DIS_FORCE_SEND_RESUME: EfuseField = EfuseField::new(EfuseBlock::Block0, 18, 1);
-/// `[]` wr_dis of SECURE_VERSION
-pub const WR_DIS_SECURE_VERSION: EfuseField = EfuseField::new(EfuseBlock::Block0, 18, 1);
-/// `[]` wr_dis of SECURE_BOOT_DISABLE_FAST_WAKE
-pub const WR_DIS_SECURE_BOOT_DISABLE_FAST_WAKE: EfuseField =
-    EfuseField::new(EfuseBlock::Block0, 18, 1);
-/// `[]` wr_dis of HYS_EN_PAD0
-pub const WR_DIS_HYS_EN_PAD0: EfuseField = EfuseField::new(EfuseBlock::Block0, 19, 1);
-/// `[]` wr_dis of HYS_EN_PAD1
-pub const WR_DIS_HYS_EN_PAD1: EfuseField = EfuseField::new(EfuseBlock::Block0, 19, 1);
-/// `[]` wr_dis of BLOCK1
-pub const WR_DIS_BLK1: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[WR_DIS.MAC_FACTORY]` wr_dis of MAC
-pub const WR_DIS_MAC: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of MAC_EXT
-pub const WR_DIS_MAC_EXT: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of RXIQ_VERSION
-pub const WR_DIS_RXIQ_VERSION: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of RXIQ_0
-pub const WR_DIS_RXIQ_0: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of RXIQ_1
-pub const WR_DIS_RXIQ_1: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of ACTIVE_HP_DBIAS
-pub const WR_DIS_ACTIVE_HP_DBIAS: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of ACTIVE_LP_DBIAS
-pub const WR_DIS_ACTIVE_LP_DBIAS: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of DSLP_DBIAS
-pub const WR_DIS_DSLP_DBIAS: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of DBIAS_VOL_GAP
-pub const WR_DIS_DBIAS_VOL_GAP: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of WAFER_VERSION_MINOR
-pub const WR_DIS_WAFER_VERSION_MINOR: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of WAFER_VERSION_MAJOR
-pub const WR_DIS_WAFER_VERSION_MAJOR: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of DISABLE_WAFER_VERSION_MAJOR
-pub const WR_DIS_DISABLE_WAFER_VERSION_MAJOR: EfuseField =
-    EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of FLASH_CAP
-pub const WR_DIS_FLASH_CAP: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of FLASH_TEMP
-pub const WR_DIS_FLASH_TEMP: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of FLASH_VENDOR
-pub const WR_DIS_FLASH_VENDOR: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of PKG_VERSION
-pub const WR_DIS_PKG_VERSION: EfuseField = EfuseField::new(EfuseBlock::Block0, 20, 1);
-/// `[]` wr_dis of BLOCK2
-pub const WR_DIS_SYS_DATA_PART1: EfuseField = EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of OPTIONAL_UNIQUE_ID
-pub const WR_DIS_OPTIONAL_UNIQUE_ID: EfuseField = EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of BLK_VERSION_MINOR
-pub const WR_DIS_BLK_VERSION_MINOR: EfuseField = EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of BLK_VERSION_MAJOR
-pub const WR_DIS_BLK_VERSION_MAJOR: EfuseField = EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of DISABLE_BLK_VERSION_MAJOR
-pub const WR_DIS_DISABLE_BLK_VERSION_MAJOR: EfuseField = EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of TEMP_CALIB
-pub const WR_DIS_TEMP_CALIB: EfuseField = EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of ADC1_AVE_INITCODE_ATTEN0
-pub const WR_DIS_ADC1_AVE_INITCODE_ATTEN0: EfuseField = EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of ADC1_AVE_INITCODE_ATTEN1
-pub const WR_DIS_ADC1_AVE_INITCODE_ATTEN1: EfuseField = EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of ADC1_AVE_INITCODE_ATTEN2
-pub const WR_DIS_ADC1_AVE_INITCODE_ATTEN2: EfuseField = EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of ADC1_AVE_INITCODE_ATTEN3
-pub const WR_DIS_ADC1_AVE_INITCODE_ATTEN3: EfuseField = EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of ADC1_HI_DOUT_ATTEN0
-pub const WR_DIS_ADC1_HI_DOUT_ATTEN0: EfuseField = EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of ADC1_HI_DOUT_ATTEN1
-pub const WR_DIS_ADC1_HI_DOUT_ATTEN1: EfuseField = EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of ADC1_HI_DOUT_ATTEN2
-pub const WR_DIS_ADC1_HI_DOUT_ATTEN2: EfuseField = EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of ADC1_HI_DOUT_ATTEN3
-pub const WR_DIS_ADC1_HI_DOUT_ATTEN3: EfuseField = EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of ADC1_CH0_ATTEN0_INITCODE_DIFF
-pub const WR_DIS_ADC1_CH0_ATTEN0_INITCODE_DIFF: EfuseField =
-    EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of ADC1_CH1_ATTEN0_INITCODE_DIFF
-pub const WR_DIS_ADC1_CH1_ATTEN0_INITCODE_DIFF: EfuseField =
-    EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of ADC1_CH2_ATTEN0_INITCODE_DIFF
-pub const WR_DIS_ADC1_CH2_ATTEN0_INITCODE_DIFF: EfuseField =
-    EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of ADC1_CH3_ATTEN0_INITCODE_DIFF
-pub const WR_DIS_ADC1_CH3_ATTEN0_INITCODE_DIFF: EfuseField =
-    EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[]` wr_dis of ADC1_CH4_ATTEN0_INITCODE_DIFF
-pub const WR_DIS_ADC1_CH4_ATTEN0_INITCODE_DIFF: EfuseField =
-    EfuseField::new(EfuseBlock::Block0, 21, 1);
-/// `[WR_DIS.USER_DATA]` wr_dis of BLOCK_USR_DATA
-pub const WR_DIS_BLOCK_USR_DATA: EfuseField = EfuseField::new(EfuseBlock::Block0, 22, 1);
-/// `[WR_DIS.MAC_CUSTOM WR_DIS.USER_DATA_MAC_CUSTOM]` wr_dis of CUSTOM_MAC
-pub const WR_DIS_CUSTOM_MAC: EfuseField = EfuseField::new(EfuseBlock::Block0, 22, 1);
-/// `[WR_DIS.KEY0]` wr_dis of BLOCK_KEY0
-pub const WR_DIS_BLOCK_KEY0: EfuseField = EfuseField::new(EfuseBlock::Block0, 23, 1);
-/// `[WR_DIS.KEY1]` wr_dis of BLOCK_KEY1
-pub const WR_DIS_BLOCK_KEY1: EfuseField = EfuseField::new(EfuseBlock::Block0, 24, 1);
-/// `[WR_DIS.KEY2]` wr_dis of BLOCK_KEY2
-pub const WR_DIS_BLOCK_KEY2: EfuseField = EfuseField::new(EfuseBlock::Block0, 25, 1);
-/// `[WR_DIS.KEY3]` wr_dis of BLOCK_KEY3
-pub const WR_DIS_BLOCK_KEY3: EfuseField = EfuseField::new(EfuseBlock::Block0, 26, 1);
-/// `[WR_DIS.KEY4]` wr_dis of BLOCK_KEY4
-pub const WR_DIS_BLOCK_KEY4: EfuseField = EfuseField::new(EfuseBlock::Block0, 27, 1);
-/// `[WR_DIS.KEY5]` wr_dis of BLOCK_KEY5
-pub const WR_DIS_BLOCK_KEY5: EfuseField = EfuseField::new(EfuseBlock::Block0, 28, 1);
-/// `[WR_DIS.SYS_DATA_PART2]` wr_dis of BLOCK_SYS_DATA2
-pub const WR_DIS_BLOCK_SYS_DATA2: EfuseField = EfuseField::new(EfuseBlock::Block0, 29, 1);
-/// `[]` wr_dis of USB_EXCHG_PINS
-pub const WR_DIS_USB_EXCHG_PINS: EfuseField = EfuseField::new(EfuseBlock::Block0, 30, 1);
-/// `[]` wr_dis of VDD_SPI_AS_GPIO
-pub const WR_DIS_VDD_SPI_AS_GPIO: EfuseField = EfuseField::new(EfuseBlock::Block0, 30, 1);
-/// `[]` wr_dis of SOFT_DIS_JTAG
-pub const WR_DIS_SOFT_DIS_JTAG: EfuseField = EfuseField::new(EfuseBlock::Block0, 31, 1);
-/// `[]` Disable reading from BlOCK4-10
-pub const RD_DIS: EfuseField = EfuseField::new(EfuseBlock::Block0, 32, 7);
-/// `[RD_DIS.KEY0]` rd_dis of BLOCK_KEY0
-pub const RD_DIS_BLOCK_KEY0: EfuseField = EfuseField::new(EfuseBlock::Block0, 32, 1);
-/// `[RD_DIS.KEY1]` rd_dis of BLOCK_KEY1
-pub const RD_DIS_BLOCK_KEY1: EfuseField = EfuseField::new(EfuseBlock::Block0, 33, 1);
-/// `[RD_DIS.KEY2]` rd_dis of BLOCK_KEY2
-pub const RD_DIS_BLOCK_KEY2: EfuseField = EfuseField::new(EfuseBlock::Block0, 34, 1);
-/// `[RD_DIS.KEY3]` rd_dis of BLOCK_KEY3
-pub const RD_DIS_BLOCK_KEY3: EfuseField = EfuseField::new(EfuseBlock::Block0, 35, 1);
-/// `[RD_DIS.KEY4]` rd_dis of BLOCK_KEY4
-pub const RD_DIS_BLOCK_KEY4: EfuseField = EfuseField::new(EfuseBlock::Block0, 36, 1);
-/// `[RD_DIS.KEY5]` rd_dis of BLOCK_KEY5
-pub const RD_DIS_BLOCK_KEY5: EfuseField = EfuseField::new(EfuseBlock::Block0, 37, 1);
-/// `[RD_DIS.SYS_DATA_PART2]` rd_dis of BLOCK_SYS_DATA2
-pub const RD_DIS_BLOCK_SYS_DATA2: EfuseField = EfuseField::new(EfuseBlock::Block0, 38, 1);
-/// `[]` Represents whether icache is disabled or enabled. 1: disabled. 0:
+use super::EfuseField;
+
+/// Disable programming of individual eFuses
+pub const WR_DIS: EfuseField = EfuseField::new(0, 0, 0, 32);
+/// Disable reading from BlOCK4-10
+pub const RD_DIS: EfuseField = EfuseField::new(0, 1, 32, 7);
+/// Reserved
+pub const RPT4_RESERVED0_4: EfuseField = EfuseField::new(0, 1, 39, 1);
+/// Represents whether icache is disabled or enabled. 1: disabled. 0: enabled
+pub const DIS_ICACHE: EfuseField = EfuseField::new(0, 1, 40, 1);
+/// Represents whether the function of usb switch to jtag is disabled or
+/// enabled. 1: disabled. 0: enabled
+pub const DIS_USB_JTAG: EfuseField = EfuseField::new(0, 1, 41, 1);
+/// Represents whether power glitch function is enabled. 1: enabled. 0: disabled
+pub const POWERGLITCH_EN: EfuseField = EfuseField::new(0, 1, 42, 1);
+/// Represents whether USB-Serial-JTAG is disabled or enabled. 1: disabled. 0:
 /// enabled
-pub const DIS_ICACHE: EfuseField = EfuseField::new(EfuseBlock::Block0, 40, 1);
-/// `[]` Represents whether the function of usb switch to jtag is disabled or
-/// enabled. 1: disabled. 0: enabled
-pub const DIS_USB_JTAG: EfuseField = EfuseField::new(EfuseBlock::Block0, 41, 1);
-/// `[]` Represents whether power glitch function is enabled. 1: enabled. 0:
-/// disabled
-pub const POWERGLITCH_EN: EfuseField = EfuseField::new(EfuseBlock::Block0, 42, 1);
-/// `[]` Represents whether the function that forces chip into download mode is
+pub const DIS_USB_SERIAL_JTAG: EfuseField = EfuseField::new(0, 1, 43, 1);
+/// Represents whether the function that forces chip into download mode is
 /// disabled or enabled. 1: disabled. 0: enabled
-pub const DIS_FORCE_DOWNLOAD: EfuseField = EfuseField::new(EfuseBlock::Block0, 44, 1);
-/// `[]` Represents whether SPI0 controller during boot_mode_download is
-/// disabled or enabled. 1: disabled. 0: enabled
-pub const SPI_DOWNLOAD_MSPI_DIS: EfuseField = EfuseField::new(EfuseBlock::Block0, 45, 1);
-/// `[DIS_CAN]` Represents whether TWAI function is disabled or enabled. 1:
-/// disabled. 0: enabled
-pub const DIS_TWAI: EfuseField = EfuseField::new(EfuseBlock::Block0, 46, 1);
-/// `[]` Set this bit to enable selection between usb_to_jtag and pad_to_jtag
-/// through strapping gpio25 when both EFUSE_DIS_PAD_JTAG and EFUSE_DIS_USB_JTAG
-/// are equal to 0
-pub const JTAG_SEL_ENABLE: EfuseField = EfuseField::new(EfuseBlock::Block0, 47, 1);
-/// `[]` Represents whether JTAG is disabled in soft way. Odd number: disabled.
-/// Even number: enabled
-pub const SOFT_DIS_JTAG: EfuseField = EfuseField::new(EfuseBlock::Block0, 48, 3);
-/// `[]` Represents whether JTAG is disabled in the hard way(permanently). 1:
-/// disabled. 0: enabled
-pub const DIS_PAD_JTAG: EfuseField = EfuseField::new(EfuseBlock::Block0, 51, 1);
-/// `[]` Represents whether flash encrypt function is disabled or enabled(except
-/// in SPI boot mode). 1: disabled. 0: enabled
-pub const DIS_DOWNLOAD_MANUAL_ENCRYPT: EfuseField = EfuseField::new(EfuseBlock::Block0, 52, 1);
-/// `[]` Represents whether the D+ and D- pins is exchanged. 1: exchanged. 0:
-/// not exchanged
-pub const USB_EXCHG_PINS: EfuseField = EfuseField::new(EfuseBlock::Block0, 57, 1);
-/// `[]` Represents whether vdd spi pin is functioned as gpio. 1: functioned. 0:
-/// not functioned
-pub const VDD_SPI_AS_GPIO: EfuseField = EfuseField::new(EfuseBlock::Block0, 58, 1);
-/// `[]` Represents whether RTC watchdog timeout threshold is selected at
-/// startup. 1: selected. 0: not selected
-pub const WDT_DELAY_SEL: EfuseField = EfuseField::new(EfuseBlock::Block0, 80, 2);
-/// `[]` Enables flash encryption when 1 or 3 bits are set and disables
-/// otherwise {0: "Disable"; 1: "Enable"; 3: "Disable"; 7: "Enable"}
-pub const SPI_BOOT_CRYPT_CNT: EfuseField = EfuseField::new(EfuseBlock::Block0, 82, 3);
-/// `[]` Revoke 1st secure boot key
-pub const SECURE_BOOT_KEY_REVOKE0: EfuseField = EfuseField::new(EfuseBlock::Block0, 85, 1);
-/// `[]` Revoke 2nd secure boot key
-pub const SECURE_BOOT_KEY_REVOKE1: EfuseField = EfuseField::new(EfuseBlock::Block0, 86, 1);
-/// `[]` Revoke 3rd secure boot key
-pub const SECURE_BOOT_KEY_REVOKE2: EfuseField = EfuseField::new(EfuseBlock::Block0, 87, 1);
-/// `[KEY0_PURPOSE]` Represents the purpose of Key0
-pub const KEY_PURPOSE_0: EfuseField = EfuseField::new(EfuseBlock::Block0, 88, 4);
-/// `[KEY1_PURPOSE]` Represents the purpose of Key1
-pub const KEY_PURPOSE_1: EfuseField = EfuseField::new(EfuseBlock::Block0, 92, 4);
-/// `[KEY2_PURPOSE]` Represents the purpose of Key2
-pub const KEY_PURPOSE_2: EfuseField = EfuseField::new(EfuseBlock::Block0, 96, 4);
-/// `[KEY3_PURPOSE]` Represents the purpose of Key3
-pub const KEY_PURPOSE_3: EfuseField = EfuseField::new(EfuseBlock::Block0, 100, 4);
-/// `[KEY4_PURPOSE]` Represents the purpose of Key4
-pub const KEY_PURPOSE_4: EfuseField = EfuseField::new(EfuseBlock::Block0, 104, 4);
-/// `[KEY5_PURPOSE]` Represents the purpose of Key5
-pub const KEY_PURPOSE_5: EfuseField = EfuseField::new(EfuseBlock::Block0, 108, 4);
-/// `[]` Represents the spa secure level by configuring the clock random divide
-/// mode
-pub const SEC_DPA_LEVEL: EfuseField = EfuseField::new(EfuseBlock::Block0, 112, 2);
-/// `[]` Represents whether hardware random number k is forced used in ESDCA. 1:
-/// force used. 0: not force used
-pub const ECDSA_FORCE_USE_HARDWARE_K: EfuseField = EfuseField::new(EfuseBlock::Block0, 114, 1);
-/// `[]` Represents whether anti-dpa attack is enabled. 1:enabled. 0: disabled
-pub const CRYPT_DPA_ENABLE: EfuseField = EfuseField::new(EfuseBlock::Block0, 115, 1);
-/// `[]` Represents whether secure boot is enabled or disabled. 1: enabled. 0:
-/// disabled
-pub const SECURE_BOOT_EN: EfuseField = EfuseField::new(EfuseBlock::Block0, 116, 1);
-/// `[]` Represents whether revoking aggressive secure boot is enabled or
-/// disabled. 1: enabled. 0: disabled
-pub const SECURE_BOOT_AGGRESSIVE_REVOKE: EfuseField = EfuseField::new(EfuseBlock::Block0, 117, 1);
-/// `[]` Represents the flash waiting time after power-up; in unit of ms. When
-/// the value less than 15; the waiting time is the programmed value. Otherwise;
-/// the waiting time is 2 times the programmed value
-pub const FLASH_TPUW: EfuseField = EfuseField::new(EfuseBlock::Block0, 124, 4);
-/// `[]` Represents whether Download mode is disabled or enabled. 1: disabled.
-/// 0: enabled
-pub const DIS_DOWNLOAD_MODE: EfuseField = EfuseField::new(EfuseBlock::Block0, 128, 1);
-/// `[]` Represents whether direct boot mode is disabled or enabled. 1:
-/// disabled. 0: enabled
-pub const DIS_DIRECT_BOOT: EfuseField = EfuseField::new(EfuseBlock::Block0, 129, 1);
-/// `[DIS_USB_PRINT]` Set this bit to disable USB-Serial-JTAG print during rom
-/// boot
-pub const DIS_USB_SERIAL_JTAG_ROM_PRINT: EfuseField = EfuseField::new(EfuseBlock::Block0, 130, 1);
-/// `[]` Represents whether the USB-Serial-JTAG download function is disabled or
+pub const DIS_FORCE_DOWNLOAD: EfuseField = EfuseField::new(0, 1, 44, 1);
+/// Represents whether SPI0 controller during boot_mode_download is disabled or
 /// enabled. 1: disabled. 0: enabled
-pub const DIS_USB_SERIAL_JTAG_DOWNLOAD_MODE: EfuseField =
-    EfuseField::new(EfuseBlock::Block0, 132, 1);
-/// `[]` Represents whether security download is enabled or disabled. 1:
-/// enabled. 0: disabled
-pub const ENABLE_SECURITY_DOWNLOAD: EfuseField = EfuseField::new(EfuseBlock::Block0, 133, 1);
-/// `[]` Set the default UARTboot message output mode {0: "Enable"; 1: "Enable
-/// when GPIO8 is low at reset"; 2: "Enable when GPIO8 is high at reset"; 3:
-/// "Disable"}
-pub const UART_PRINT_CONTROL: EfuseField = EfuseField::new(EfuseBlock::Block0, 134, 2);
-/// `[]` Represents whether ROM code is forced to send a resume command during
-/// SPI boot. 1: forced. 0:not forced
-pub const FORCE_SEND_RESUME: EfuseField = EfuseField::new(EfuseBlock::Block0, 136, 1);
-/// `[]` Represents the version used by ESP-IDF anti-rollback feature
-pub const SECURE_VERSION: EfuseField = EfuseField::new(EfuseBlock::Block0, 137, 16);
-/// `[]` Represents whether FAST VERIFY ON WAKE is disabled or enabled when
-/// Secure Boot is enabled. 1: disabled. 0: enabled
-pub const SECURE_BOOT_DISABLE_FAST_WAKE: EfuseField = EfuseField::new(EfuseBlock::Block0, 153, 1);
-/// `[]` Set bits to enable hysteresis function of PAD0~5
-pub const HYS_EN_PAD0: EfuseField = EfuseField::new(EfuseBlock::Block0, 154, 6);
-/// `[]` Set bits to enable hysteresis function of PAD6~27
-pub const HYS_EN_PAD1: EfuseField = EfuseField::new(EfuseBlock::Block0, 160, 22);
-/// `[MAC_FACTORY]` MAC address
-pub const MAC: EfuseField = EfuseField::new(EfuseBlock::Block1, 0, 48);
-/// `[]` Stores the extended bits of MAC address
-pub const MAC_EXT: EfuseField = EfuseField::new(EfuseBlock::Block1, 48, 16);
-/// `[]` Stores RF Calibration data. RXIQ version
-pub const RXIQ_VERSION: EfuseField = EfuseField::new(EfuseBlock::Block1, 64, 3);
-/// `[]` Stores RF Calibration data. RXIQ data 0
-pub const RXIQ_0: EfuseField = EfuseField::new(EfuseBlock::Block1, 67, 7);
-/// `[]` Stores RF Calibration data. RXIQ data 1
-pub const RXIQ_1: EfuseField = EfuseField::new(EfuseBlock::Block1, 74, 7);
-/// `[]` Stores the PMU active hp dbias
-pub const ACTIVE_HP_DBIAS: EfuseField = EfuseField::new(EfuseBlock::Block1, 81, 5);
-/// `[]` Stores the PMU active lp dbias
-pub const ACTIVE_LP_DBIAS: EfuseField = EfuseField::new(EfuseBlock::Block1, 86, 5);
-/// `[]` Stores the PMU sleep dbias
-pub const DSLP_DBIAS: EfuseField = EfuseField::new(EfuseBlock::Block1, 91, 4);
-/// `[]` Stores the low 1 bit of dbias_vol_gap
-pub const DBIAS_VOL_GAP: EfuseField = EfuseField::new(EfuseBlock::Block1, 95, 5);
-/// `[]` Stores the wafer version minor
-pub const WAFER_VERSION_MINOR: EfuseField = EfuseField::new(EfuseBlock::Block1, 114, 3);
-/// `[]` Stores the wafer version major
-pub const WAFER_VERSION_MAJOR: EfuseField = EfuseField::new(EfuseBlock::Block1, 117, 2);
-/// `[]` Disables check of wafer version major
-pub const DISABLE_WAFER_VERSION_MAJOR: EfuseField = EfuseField::new(EfuseBlock::Block1, 119, 1);
-/// `[]` Stores the flash cap
-pub const FLASH_CAP: EfuseField = EfuseField::new(EfuseBlock::Block1, 120, 3);
-/// `[]` Stores the flash temp
-pub const FLASH_TEMP: EfuseField = EfuseField::new(EfuseBlock::Block1, 123, 2);
-/// `[]` Stores the flash vendor
-pub const FLASH_VENDOR: EfuseField = EfuseField::new(EfuseBlock::Block1, 125, 3);
-/// `[]` Package version
-pub const PKG_VERSION: EfuseField = EfuseField::new(EfuseBlock::Block1, 128, 3);
-/// `[]` Optional unique 128-bit ID
-pub const OPTIONAL_UNIQUE_ID: EfuseField = EfuseField::new(EfuseBlock::Block2, 0, 128);
-/// `[]` BLK_VERSION_MINOR of BLOCK2. 1: RF Calibration data in BLOCK1
-pub const BLK_VERSION_MINOR: EfuseField = EfuseField::new(EfuseBlock::Block2, 130, 3);
-/// `[]` BLK_VERSION_MAJOR of BLOCK2
-pub const BLK_VERSION_MAJOR: EfuseField = EfuseField::new(EfuseBlock::Block2, 133, 2);
-/// `[]` Disables check of blk version major
-pub const DISABLE_BLK_VERSION_MAJOR: EfuseField = EfuseField::new(EfuseBlock::Block2, 135, 1);
-/// `[]` Temperature calibration data
-pub const TEMP_CALIB: EfuseField = EfuseField::new(EfuseBlock::Block2, 136, 9);
-/// `[]` ADC1 calibration data
-pub const ADC1_AVE_INITCODE_ATTEN0: EfuseField = EfuseField::new(EfuseBlock::Block2, 145, 10);
-/// `[]` ADC1 calibration data
-pub const ADC1_AVE_INITCODE_ATTEN1: EfuseField = EfuseField::new(EfuseBlock::Block2, 155, 10);
-/// `[]` ADC1 calibration data
-pub const ADC1_AVE_INITCODE_ATTEN2: EfuseField = EfuseField::new(EfuseBlock::Block2, 165, 10);
-/// `[]` ADC1 calibration data
-pub const ADC1_AVE_INITCODE_ATTEN3: EfuseField = EfuseField::new(EfuseBlock::Block2, 175, 10);
-/// `[]` ADC1 calibration data
-pub const ADC1_HI_DOUT_ATTEN0: EfuseField = EfuseField::new(EfuseBlock::Block2, 185, 10);
-/// `[]` ADC1 calibration data
-pub const ADC1_HI_DOUT_ATTEN1: EfuseField = EfuseField::new(EfuseBlock::Block2, 195, 10);
-/// `[]` ADC1 calibration data
-pub const ADC1_HI_DOUT_ATTEN2: EfuseField = EfuseField::new(EfuseBlock::Block2, 205, 10);
-/// `[]` ADC1 calibration data
-pub const ADC1_HI_DOUT_ATTEN3: EfuseField = EfuseField::new(EfuseBlock::Block2, 215, 10);
-/// `[]` ADC1 calibration data
-pub const ADC1_CH0_ATTEN0_INITCODE_DIFF: EfuseField = EfuseField::new(EfuseBlock::Block2, 225, 4);
-/// `[]` ADC1 calibration data
-pub const ADC1_CH1_ATTEN0_INITCODE_DIFF: EfuseField = EfuseField::new(EfuseBlock::Block2, 229, 4);
-/// `[]` ADC1 calibration data
-pub const ADC1_CH2_ATTEN0_INITCODE_DIFF: EfuseField = EfuseField::new(EfuseBlock::Block2, 233, 4);
-/// `[]` ADC1 calibration data
-pub const ADC1_CH3_ATTEN0_INITCODE_DIFF: EfuseField = EfuseField::new(EfuseBlock::Block2, 237, 4);
-/// `[]` ADC1 calibration data
-pub const ADC1_CH4_ATTEN0_INITCODE_DIFF: EfuseField = EfuseField::new(EfuseBlock::Block2, 241, 4);
-/// `[BLOCK_USR_DATA]` User data
-pub const USER_DATA: EfuseField = EfuseField::new(EfuseBlock::Block3, 0, 256);
-/// `[MAC_CUSTOM CUSTOM_MAC]` Custom MAC
-pub const USER_DATA_MAC_CUSTOM: EfuseField = EfuseField::new(EfuseBlock::Block3, 200, 48);
-/// `[BLOCK_KEY0]` Key0 or user data
-pub const KEY0: EfuseField = EfuseField::new(EfuseBlock::Block4, 0, 256);
-/// `[BLOCK_KEY1]` Key1 or user data
-pub const KEY1: EfuseField = EfuseField::new(EfuseBlock::Block5, 0, 256);
-/// `[BLOCK_KEY2]` Key2 or user data
-pub const KEY2: EfuseField = EfuseField::new(EfuseBlock::Block6, 0, 256);
-/// `[BLOCK_KEY3]` Key3 or user data
-pub const KEY3: EfuseField = EfuseField::new(EfuseBlock::Block7, 0, 256);
-/// `[BLOCK_KEY4]` Key4 or user data
-pub const KEY4: EfuseField = EfuseField::new(EfuseBlock::Block8, 0, 256);
-/// `[BLOCK_KEY5]` Key5 or user data
-pub const KEY5: EfuseField = EfuseField::new(EfuseBlock::Block9, 0, 256);
-/// `[BLOCK_SYS_DATA2]` System data part 2 (reserved)
-pub const SYS_DATA_PART2: EfuseField = EfuseField::new(EfuseBlock::Block10, 0, 256);
+pub const SPI_DOWNLOAD_MSPI_DIS: EfuseField = EfuseField::new(0, 1, 45, 1);
+/// Represents whether TWAI function is disabled or enabled. 1: disabled. 0:
+/// enabled
+pub const DIS_TWAI: EfuseField = EfuseField::new(0, 1, 46, 1);
+/// Set this bit to enable selection between usb_to_jtag and pad_to_jtag through
+/// strapping gpio25 when both EFUSE_DIS_PAD_JTAG and EFUSE_DIS_USB_JTAG are
+/// equal to 0
+pub const JTAG_SEL_ENABLE: EfuseField = EfuseField::new(0, 1, 47, 1);
+/// Represents whether JTAG is disabled in soft way. Odd number: disabled. Even
+/// number: enabled
+pub const SOFT_DIS_JTAG: EfuseField = EfuseField::new(0, 1, 48, 3);
+/// Represents whether JTAG is disabled in the hard way(permanently). 1:
+/// disabled. 0: enabled
+pub const DIS_PAD_JTAG: EfuseField = EfuseField::new(0, 1, 51, 1);
+/// Represents whether flash encrypt function is disabled or enabled(except in
+/// SPI boot mode). 1: disabled. 0: enabled
+pub const DIS_DOWNLOAD_MANUAL_ENCRYPT: EfuseField = EfuseField::new(0, 1, 52, 1);
+/// Represents the single-end input threshold vrefh; 1.76 V to 2 V with step of
+/// 80 mV
+pub const USB_DREFH: EfuseField = EfuseField::new(0, 1, 53, 2);
+/// Represents the single-end input threshold vrefl; 1.76 V to 2 V with step of
+/// 80 mV
+pub const USB_DREFL: EfuseField = EfuseField::new(0, 1, 55, 2);
+/// Represents whether the D+ and D- pins is exchanged. 1: exchanged. 0: not
+/// exchanged
+pub const USB_EXCHG_PINS: EfuseField = EfuseField::new(0, 1, 57, 1);
+/// Represents whether vdd spi pin is functioned as gpio. 1: functioned. 0: not
+/// functioned
+pub const VDD_SPI_AS_GPIO: EfuseField = EfuseField::new(0, 1, 58, 1);
+/// Configures the curve of ECDSA calculation: 0: only enable P256. 1: only
+/// enable P192. 2: both enable P256 and P192. 3: only enable P256
+pub const ECDSA_CURVE_MODE: EfuseField = EfuseField::new(0, 1, 59, 2);
+/// Set this bit to permanently turn on ECC const-time mode
+pub const ECC_FORCE_CONST_TIME: EfuseField = EfuseField::new(0, 1, 61, 1);
+/// Set this bit to control the xts pseudo-round anti-dpa attack function: 0:
+/// controlled by register. 1-3: the higher the value is; the more pseudo-rounds
+/// are inserted to the xts-aes calculation
+pub const XTS_DPA_PSEUDO_LEVEL: EfuseField = EfuseField::new(0, 1, 62, 2);
+/// Reserved
+pub const RPT4_RESERVED1_1: EfuseField = EfuseField::new(0, 2, 64, 16);
+/// Represents whether RTC watchdog timeout threshold is selected at startup. 1:
+/// selected. 0: not selected
+pub const WDT_DELAY_SEL: EfuseField = EfuseField::new(0, 2, 80, 2);
+/// Enables flash encryption when 1 or 3 bits are set and disables otherwise
+pub const SPI_BOOT_CRYPT_CNT: EfuseField = EfuseField::new(0, 2, 82, 3);
+/// Revoke 1st secure boot key
+pub const SECURE_BOOT_KEY_REVOKE0: EfuseField = EfuseField::new(0, 2, 85, 1);
+/// Revoke 2nd secure boot key
+pub const SECURE_BOOT_KEY_REVOKE1: EfuseField = EfuseField::new(0, 2, 86, 1);
+/// Revoke 3rd secure boot key
+pub const SECURE_BOOT_KEY_REVOKE2: EfuseField = EfuseField::new(0, 2, 87, 1);
+/// Represents the purpose of Key0
+pub const KEY_PURPOSE_0: EfuseField = EfuseField::new(0, 2, 88, 4);
+/// Represents the purpose of Key1
+pub const KEY_PURPOSE_1: EfuseField = EfuseField::new(0, 2, 92, 4);
+/// Represents the purpose of Key2
+pub const KEY_PURPOSE_2: EfuseField = EfuseField::new(0, 3, 96, 4);
+/// Represents the purpose of Key3
+pub const KEY_PURPOSE_3: EfuseField = EfuseField::new(0, 3, 100, 4);
+/// Represents the purpose of Key4
+pub const KEY_PURPOSE_4: EfuseField = EfuseField::new(0, 3, 104, 4);
+/// Represents the purpose of Key5
+pub const KEY_PURPOSE_5: EfuseField = EfuseField::new(0, 3, 108, 4);
+/// Represents the spa secure level by configuring the clock random divide mode
+pub const SEC_DPA_LEVEL: EfuseField = EfuseField::new(0, 3, 112, 2);
+/// Reserved
+pub const RESERVE_0_114: EfuseField = EfuseField::new(0, 3, 114, 1);
+/// Represents whether anti-dpa attack is enabled. 1:enabled. 0: disabled
+pub const CRYPT_DPA_ENABLE: EfuseField = EfuseField::new(0, 3, 115, 1);
+/// Represents whether secure boot is enabled or disabled. 1: enabled. 0:
+/// disabled
+pub const SECURE_BOOT_EN: EfuseField = EfuseField::new(0, 3, 116, 1);
+/// Represents whether revoking aggressive secure boot is enabled or disabled.
+/// 1: enabled. 0: disabled
+pub const SECURE_BOOT_AGGRESSIVE_REVOKE: EfuseField = EfuseField::new(0, 3, 117, 1);
+/// Set these bits to enable power glitch function when chip power on
+pub const POWERGLITCH_EN1: EfuseField = EfuseField::new(0, 3, 118, 5);
+/// reserved
+pub const RESERVED_0_123: EfuseField = EfuseField::new(0, 3, 123, 1);
+/// Represents the flash waiting time after power-up; in unit of ms. When the
+/// value less than 15; the waiting time is the programmed value. Otherwise; the
+/// waiting time is 2 times the programmed value
+pub const FLASH_TPUW: EfuseField = EfuseField::new(0, 3, 124, 4);
+/// Represents whether Download mode is disabled or enabled. 1: disabled. 0:
+/// enabled
+pub const DIS_DOWNLOAD_MODE: EfuseField = EfuseField::new(0, 4, 128, 1);
+/// Represents whether direct boot mode is disabled or enabled. 1: disabled. 0:
+/// enabled
+pub const DIS_DIRECT_BOOT: EfuseField = EfuseField::new(0, 4, 129, 1);
+/// Set this bit to disable USB-Serial-JTAG print during rom boot
+pub const DIS_USB_SERIAL_JTAG_ROM_PRINT: EfuseField = EfuseField::new(0, 4, 130, 1);
+/// Reserved
+pub const RPT4_RESERVED3_5: EfuseField = EfuseField::new(0, 4, 131, 1);
+/// Represents whether the USB-Serial-JTAG download function is disabled or
+/// enabled. 1: disabled. 0: enabled
+pub const DIS_USB_SERIAL_JTAG_DOWNLOAD_MODE: EfuseField = EfuseField::new(0, 4, 132, 1);
+/// Represents whether security download is enabled or disabled. 1: enabled. 0:
+/// disabled
+pub const ENABLE_SECURITY_DOWNLOAD: EfuseField = EfuseField::new(0, 4, 133, 1);
+/// Set the default UARTboot message output mode
+pub const UART_PRINT_CONTROL: EfuseField = EfuseField::new(0, 4, 134, 2);
+/// Represents whether ROM code is forced to send a resume command during SPI
+/// boot. 1: forced. 0:not forced
+pub const FORCE_SEND_RESUME: EfuseField = EfuseField::new(0, 4, 136, 1);
+/// Represents the version used by ESP-IDF anti-rollback feature
+pub const SECURE_VERSION: EfuseField = EfuseField::new(0, 4, 137, 16);
+/// Represents whether FAST VERIFY ON WAKE is disabled or enabled when Secure
+/// Boot is enabled. 1: disabled. 0: enabled
+pub const SECURE_BOOT_DISABLE_FAST_WAKE: EfuseField = EfuseField::new(0, 4, 153, 1);
+/// Set bits to enable hysteresis function of PAD0~5
+pub const HYS_EN_PAD0: EfuseField = EfuseField::new(0, 4, 154, 6);
+/// Set bits to enable hysteresis function of PAD6~27
+pub const HYS_EN_PAD1: EfuseField = EfuseField::new(0, 5, 160, 22);
+/// Reserved
+pub const RPT4_RESERVED4_1: EfuseField = EfuseField::new(0, 5, 182, 2);
+/// Reserved
+pub const RPT4_RESERVED4_0: EfuseField = EfuseField::new(0, 5, 184, 8);
+/// MAC address
+pub const MAC0: EfuseField = EfuseField::new(1, 0, 0, 32);
+/// MAC address
+pub const MAC1: EfuseField = EfuseField::new(1, 1, 32, 16);
+/// Stores the extended bits of MAC address
+pub const MAC_EXT: EfuseField = EfuseField::new(1, 1, 48, 16);
+/// Stores RF Calibration data. RXIQ version
+pub const RXIQ_VERSION: EfuseField = EfuseField::new(1, 2, 64, 3);
+/// Stores RF Calibration data. RXIQ data 0
+pub const RXIQ_0: EfuseField = EfuseField::new(1, 2, 67, 7);
+/// Stores RF Calibration data. RXIQ data 1
+pub const RXIQ_1: EfuseField = EfuseField::new(1, 2, 74, 7);
+/// Stores the PMU active hp dbias
+pub const ACTIVE_HP_DBIAS: EfuseField = EfuseField::new(1, 2, 81, 5);
+/// Stores the PMU active lp dbias
+pub const ACTIVE_LP_DBIAS: EfuseField = EfuseField::new(1, 2, 86, 5);
+/// Stores the PMU sleep dbias
+pub const DSLP_DBIAS: EfuseField = EfuseField::new(1, 2, 91, 4);
+/// Stores the low 1 bit of dbias_vol_gap
+pub const DBIAS_VOL_GAP: EfuseField = EfuseField::new(1, 2, 95, 5);
+/// Reserved
+pub const MAC_RESERVED_2: EfuseField = EfuseField::new(1, 3, 100, 14);
+/// Stores the wafer version minor
+pub const WAFER_VERSION_MINOR: EfuseField = EfuseField::new(1, 3, 114, 3);
+/// Stores the wafer version major
+pub const WAFER_VERSION_MAJOR: EfuseField = EfuseField::new(1, 3, 117, 2);
+/// Disables check of wafer version major
+pub const DISABLE_WAFER_VERSION_MAJOR: EfuseField = EfuseField::new(1, 3, 119, 1);
+/// Stores the flash cap
+pub const FLASH_CAP: EfuseField = EfuseField::new(1, 3, 120, 3);
+/// Stores the flash temp
+pub const FLASH_TEMP: EfuseField = EfuseField::new(1, 3, 123, 2);
+/// Stores the flash vendor
+pub const FLASH_VENDOR: EfuseField = EfuseField::new(1, 3, 125, 3);
+/// Package version
+pub const PKG_VERSION: EfuseField = EfuseField::new(1, 4, 128, 3);
+/// reserved
+pub const RESERVED_1_131: EfuseField = EfuseField::new(1, 4, 131, 29);
+/// Stores the second 32 bits of the zeroth part of system data
+pub const SYS_DATA_PART0_2: EfuseField = EfuseField::new(1, 5, 160, 32);
+/// Optional unique 128-bit ID
+pub const OPTIONAL_UNIQUE_ID: EfuseField = EfuseField::new(2, 0, 0, 128);
+/// reserved
+pub const RESERVED_2_128: EfuseField = EfuseField::new(2, 4, 128, 2);
+/// BLK_VERSION_MINOR of BLOCK2. 1: RF Calibration data in BLOCK1
+pub const BLK_VERSION_MINOR: EfuseField = EfuseField::new(2, 4, 130, 3);
+/// BLK_VERSION_MAJOR of BLOCK2
+pub const BLK_VERSION_MAJOR: EfuseField = EfuseField::new(2, 4, 133, 2);
+/// Disables check of blk version major
+pub const DISABLE_BLK_VERSION_MAJOR: EfuseField = EfuseField::new(2, 4, 135, 1);
+/// Temperature calibration data
+pub const TEMP_CALIB: EfuseField = EfuseField::new(2, 4, 136, 9);
+/// ADC1 calibration data
+pub const ADC1_AVE_INITCODE_ATTEN0: EfuseField = EfuseField::new(2, 4, 145, 10);
+/// ADC1 calibration data
+pub const ADC1_AVE_INITCODE_ATTEN1: EfuseField = EfuseField::new(2, 4, 155, 10);
+/// ADC1 calibration data
+pub const ADC1_AVE_INITCODE_ATTEN2: EfuseField = EfuseField::new(2, 5, 165, 10);
+/// ADC1 calibration data
+pub const ADC1_AVE_INITCODE_ATTEN3: EfuseField = EfuseField::new(2, 5, 175, 10);
+/// ADC1 calibration data
+pub const ADC1_HI_DOUT_ATTEN0: EfuseField = EfuseField::new(2, 5, 185, 10);
+/// ADC1 calibration data
+pub const ADC1_HI_DOUT_ATTEN1: EfuseField = EfuseField::new(2, 6, 195, 10);
+/// ADC1 calibration data
+pub const ADC1_HI_DOUT_ATTEN2: EfuseField = EfuseField::new(2, 6, 205, 10);
+/// ADC1 calibration data
+pub const ADC1_HI_DOUT_ATTEN3: EfuseField = EfuseField::new(2, 6, 215, 10);
+/// ADC1 calibration data
+pub const ADC1_CH0_ATTEN0_INITCODE_DIFF: EfuseField = EfuseField::new(2, 7, 225, 4);
+/// ADC1 calibration data
+pub const ADC1_CH1_ATTEN0_INITCODE_DIFF: EfuseField = EfuseField::new(2, 7, 229, 4);
+/// ADC1 calibration data
+pub const ADC1_CH2_ATTEN0_INITCODE_DIFF: EfuseField = EfuseField::new(2, 7, 233, 4);
+/// ADC1 calibration data
+pub const ADC1_CH3_ATTEN0_INITCODE_DIFF: EfuseField = EfuseField::new(2, 7, 237, 4);
+/// ADC1 calibration data
+pub const ADC1_CH4_ATTEN0_INITCODE_DIFF: EfuseField = EfuseField::new(2, 7, 241, 4);
+/// reserved
+pub const RESERVED_2_245: EfuseField = EfuseField::new(2, 7, 245, 11);
+/// User data
+pub const BLOCK_USR_DATA: EfuseField = EfuseField::new(3, 0, 0, 192);
+/// reserved
+pub const RESERVED_3_192: EfuseField = EfuseField::new(3, 6, 192, 8);
+/// Custom MAC
+pub const CUSTOM_MAC: EfuseField = EfuseField::new(3, 6, 200, 48);
+/// reserved
+pub const RESERVED_3_248: EfuseField = EfuseField::new(3, 7, 248, 8);
+/// Key0 or user data
+pub const BLOCK_KEY0: EfuseField = EfuseField::new(4, 0, 0, 256);
+/// Key1 or user data
+pub const BLOCK_KEY1: EfuseField = EfuseField::new(5, 0, 0, 256);
+/// Key2 or user data
+pub const BLOCK_KEY2: EfuseField = EfuseField::new(6, 0, 0, 256);
+/// Key3 or user data
+pub const BLOCK_KEY3: EfuseField = EfuseField::new(7, 0, 0, 256);
+/// Key4 or user data
+pub const BLOCK_KEY4: EfuseField = EfuseField::new(8, 0, 0, 256);
+/// Key5 or user data
+pub const BLOCK_KEY5: EfuseField = EfuseField::new(9, 0, 0, 256);
+/// System data part 2 (reserved)
+pub const BLOCK_SYS_DATA2: EfuseField = EfuseField::new(10, 0, 0, 256);