+++ esp-hal/src/soc/esp32h2/efuse/mod.rs
@@ -42,7 +42,7 @@
 //! ```
 
 pub use self::fields::*;
-use crate::{analog::adc::Attenuation, peripherals::EFUSE};
+use crate::{analog::adc::Attenuation, peripherals::EFUSE, soc::efuse_field::EfuseField};
 
 mod fields;
 
@@ -50,11 +50,6 @@ mod fields;
 pub struct Efuse;
 
 impl Efuse {
-    /// Reads chip's MAC address from the eFuse storage.
-    pub fn read_base_mac_address() -> [u8; 6] {
-        Self::read_field_be(MAC)
-    }
-
     /// Get status of SPI boot encryption.
     pub fn flash_encryption() -> bool {
         (Self::read_field_le::<u8>(SPI_BOOT_CRYPT_CNT).count_ones() % 2) != 0
@@ -166,7 +161,8 @@ impl Efuse {
     }
 }
 
-#[derive(Copy, Clone)]
+#[derive(Debug, Clone, Copy, strum::FromRepr)]
+#[repr(u32)]
 pub(crate) enum EfuseBlock {
     Block0,
     Block1,